Vector Support Spike
This commit introduces initial support for vector data types in the
Neo4j.Driver. Because this is a spike, the serialization and
deserialization logic is implemented in a way that is not suitable for
the final product. When any kind of list is read from the database, the
code checks if all items are of the same type, and if that type is
supported as an element type for vectors. If so, it deserializes the
list as a vector. This is a temporary solution to allow for testing and
exploration of vector data handling in Neo4j. When writing a vector, it
is serialized as a list.